+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## unreleased
+
+- Add dependencyConfiguration to mauiKmp plugin configuration
+
 ## v0.2.1
 
 - Implement nuget fetching workaround for finding matching versions